+++ packages/docs-site/CHANGELOG.md
@@ -4,6 +4,14 @@
 
 ### Patch Changes
 
+- Updated dependencies [[`5d7ee8c`](https://github.com/dmno-dev/dmno/commit/5d7ee8c43a2d65e4ae7811e62ba3d2220f9ed608), [`02bc2a6`](https://github.com/dmno-dev/dmno/commit/02bc2a63c8f5e814170c08caa40e886081c40445)]:
+  - dmno@0.0.6
+  - @dmno/astro-integration@0.0.3
+
+## null
+
+### Patch Changes
+
 - Updated dependencies [[`b4a0e09`](https://github.com/dmno-dev/dmno/commit/b4a0e0992ff8d0fb832e76b5aa70c2630b86fe3a)]:
   - dmno@0.0.5
   - @dmno/astro-integration@0.0.2